+++ PyDodo/pydodo/metrics.py
@@ -4,7 +4,7 @@ import requests
 import numpy as np
 
 from . import utils
-from .utils import construct_endpoint_url
+from .bluebird_connect import construct_endpoint_url
 from .config_param import config_param
 
 endpoint = config_param("endpoint_metrics")
