+++ PyDodo/tests/integration/test_async_calls.py
@@ -12,7 +12,7 @@ from pydodo import (
     create_aircraft,
     aircraft_position,
 )
-from pydodo.utils import ping_bluebird
+from pydodo.bluebird_connect import ping_bluebird
 
 # test if can connect to BlueBird
 bb_resp = ping_bluebird()
