+++ PyDodo/tests/integration/test_create_aircraft.py
@@ -2,7 +2,7 @@ import pytest
 from requests.exceptions import HTTPError
 
 from pydodo import create_aircraft, reset_simulation
-from pydodo.utils import ping_bluebird
+from pydodo.bluebird_connect import ping_bluebird
 
 # Valid input parameter values
 aircraft_id = "TST1001"
