+++ PyDodo/tests/integration/test_create_scenario.py
@@ -11,7 +11,7 @@ from pydodo import (
     all_positions,
     list_route,
 )
-from pydodo.utils import ping_bluebird
+from pydodo.bluebird_connect import ping_bluebird
 from pydodo.config_param import config_param
 
 bb_resp = ping_bluebird()
